provision worker+submitter ephemeral storage as loopback
After some testing, this seems to work well by using an initCointainer in order to create a sparse image and then loopback-mount it in a privileged container next to the actual master container to loopback mount it. This needs the MountPropagation kubernetes feature-gate to be enabled. We also use local/ephemeral storage limits here since all of this resides on an emptyDir (in /var/lib/kubelet on the node).