use fallocate instead of truncate
We don’t really want spars files, we want the full allocation to be visible to the filesystem on creation of the image. fallocate on Linux will do that for most file systems and is basically equivalent to zeroing out the image, just much faster (since no actual I/O is done).